requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-community>=0.0.10
python-dotenv>=1.0.0
openai>=1.12.0 